requirements.txt
langgraph
langchain
langchain-groq
supabase
tavily-python
firecrawl-py
pydantic
python-dotenv
thefuzz
python-Levenshtein
schedule
python-telegram-bot
duckduckgo-search
